+++ storage/Cargo.toml
@@ -21,3 +21,4 @@ futures = "0.3"
 tonic = "0.6.1"
 hex = "0.4"
 metrics = "0.18"
+prost = "0.9"
